Safari-611.1.10/LayoutTests/http/tests/loading/basic-auth-remove-credentials-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "<!--frame1-->" - didStartProvisionalLoadForFrame
127.0.0.1:8000 - didReceiveAuthenticationChallenge - Responding with first:first-pw
frame "<!--frame1-->" - didCommitLoadForFrame
frame "<!--frame1-->" - didFinishDocumentLoadForFrame
frame "<!--frame1-->" - didHandleOnloadEventsForFrame
main frame - didHandleOnloadEventsForFrame
frame "<!--frame1-->" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
frame "<!--frame2-->" - didStartProvisionalLoadForFrame
127.0.0.1:8000 - didReceiveAuthenticationChallenge - Responding with second:second-pw
frame "<!--frame2-->" - didCommitLoadForFrame
frame "<!--frame2-->" - didFinishDocumentLoadForFrame
frame "<!--frame2-->" - didHandleOnloadEventsForFrame
frame "<!--frame2-->" - didFinishLoadForFrame
This test verifies that we are able to remove credentials, by checking that we get an authenticaton.
challenge for a path we normally would not get a challenge for, after having removed all credentials.
The first frame's path is /loading/resources/protected-resource.php, and we should get a challenge for it.
It will be authorized with first/first-pw.
The second frame's path is identical to the first frame, and we should get a challenge for it, because credentials have been deleted.
It will be authorized with second/second-pw.


--------
Frame: '<!--frame1-->'
--------
Authenticated as user: first password: first-pw

--------
Frame: '<!--frame2-->'
--------
Authenticated as user: second password: second-pw
